Adicionando Testes

+++ src/main/java/com/pedrohnf688/api/repository/TarefaRepository.java
@@ -7,11 +7,12 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.pedrohnf688.api.modelo.Tarefa;
 
-public interface TarefaRepository extends JpaRepository<Tarefa, Integer> {
+public interface TarefaRepository extends JpaRepository<Tarefa, Long> {
 
 	@Query(value = "SELECT * FROM Tarefa t WHERE t.status = false", nativeQuery = true)
 	List<Tarefa> findByTarefaStatus();
 
 	@Query(value = "SELECT * FROM Tarefa t WHERE t.status = true", nativeQuery = true)
 	List<Tarefa> findByTarefaStatusFeitas();
+	
 }
